youtube2prompt/requirements.txt
# Core dependencies
pytube>=15.0.0
youtube-transcript-api>=1.0.3
requests>=2.32.0
python-dotenv>=1.0.0

# Web interface
streamlit>=1.31.0

# GUI dependencies
PySide6>=6.5.0
qtawesome>=1.2.0
plyer>=2.1.0
Pillow>=11.0.0  # For icon generation and image handling